Improved save file compatibility handling.
* Add an internal version counter, starting at 1.
  Said value should be increased any time compatibility is broken, i.e.
  attempting to load an old save segfaults.
* Check this value when loading a save file.
* If incompatible, prompt whether the file should be deleted.
  If yes, start a new game.

This breaks saves but seeing how the map memory change (storing item
colours) already did, I don't think this is much of a problem.

The prompt is a *huge* improvement considering the last time 
compatibility broke and I had an unfinished game, I had no idea where 
the file was saved (didn't help that the folder was hidden, either)
and had to search the source to find its location and delete it
before I could continue playing. :( 


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@504 561b6253-61b6-49b5-99ee-40376865e0bd