De-randomize starting stats.
Stats are now chosen from one of three fixed stat packages with high
points in Str, Con, and Int, respectively. I feel that Dex is currently
not relevant enough to warrant a character build of its own (though
the combat overhaul and addition of throwing and maybe untrapping would 
change that), Wis has been lumped together with Int for this purpose, 
and Cha is going to be removed anyway.

There's also a fourth option that is currently deactivated where stats
are assigned at random, and a fifth possibility could be a point-buy
system that'd only make sense if all stats are actually relevant.

A menu to pick out of these possibilities (currently chosen randomly)
would be nice, as would an option to restrict random selection to a
subset of them.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@608 561b6253-61b6-49b5-99ee-40376865e0bd