Add new potion "recovery" healing stat drains, including dizziness.
I've added this in the place of the old Cha raising potion because I'd
hoped to preserve save compatibility. For some reason that didn't work
out and I ended up increasing the save version counter, anyway, so it
could be moved below see invisible, but I figure it doesn't matter.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@613 561b6253-61b6-49b5-99ee-40376865e0bd