The green urchin and lama nobe can now cause blindness, yikes!
Whatever they may be, I figure this makes these monsters more 
interesting. Also makes the cure blindness spell more useful.
Otherwise, it is possible (if a bit risky) to outwait the effect
and shooting these monsters from a distance is also an option.

I am also considering to allow prayer at an altar to cure 
effects like this, which might introduce some strategic 
incentive to lure some kinds of monsters to the nearest altar.

Either way, tested in a real game and blindness attacks are not
nearly as crippling as I thought they would be.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@622 561b6253-61b6-49b5-99ee-40376865e0bd