Refuse to use scrolls or potions known to be cursed.
There's never a reason to do so, so don't punish the player for typos.

It might be more interesting to give the player a reason to sometimes 
risk using said items (for example by bad effects only happening with
a certain chance) but that's probably hard to balance: for most 
scrolls and potions you could safely ignore the curse and use the item 
in safe surroundings.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@627 561b6253-61b6-49b5-99ee-40376865e0bd