Prompt when choosing the autoselected first spell in the menu.
Sometimes, I pull up the spell menu and end up casting the very first
(autoselected) spell without meaning to. This is particularly annoying
if said spell happens to be alter reality.

As an intermediate measure I've now added a prompt if your first
action in the spell menu is to confirm (via Enter etc.) the first
spell. If you typed any other key before, you won't be prompted.
This is likely to still be annoying on occasion and I could restrict it
to alt only, but I'd prefer a broader solution.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@631 561b6253-61b6-49b5-99ee-40376865e0bd