Changed the flow of game time management. Made actions interruptible. Give the result of a multi-turn action after elapsing the required number of turns. Documented some more functions. Got rid of the pointless type player_eqipment_t.
Originally, multi-turn actions would return the number of turns after finishing an action. These functions now call player_make_move() when starting the action and check for a positive return value. player_make_move() uses the provided number of turns and checks if the player is under attack and offers to abort the action. When aborting the action, false is returned and the functions calling player_make_move() have to abort. 

Repaint the screen and sleep for 50 microseconds per turn during actions which consume multiple turns. This causes an unsatisfying redraw of the inventory dialog, if it is opened. Will have to investigate that later...



git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@633 561b6253-61b6-49b5-99ee-40376865e0bd