Make hell hounds and red dragons breathe bursts of fire, yay!
The damage values will need to be reviewed, as will (eventually) the 
waiting and run commands.

I've added an evasion check that depends highly on player level and
difficulty setting, and also a bit on Dex, though the importance of the
latter should probably be increased. Monsters can't evade spells at the 
moment which maybe should be changed at some point.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@637 561b6253-61b6-49b5-99ee-40376865e0bd