Tweak monsters' gold property to make gold drops less reliable.
It feels unnatural to have every instance of a monster kind drop a
highly similar amount of gold coins. I've added a new parameter, 
gold_chance, that is used to decide whether an eligible monster 
should carry gold or not. The gold amounts have been increased to
make up for the lower likelihood, so the overall gold distribution
should be roughly the same.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@641 561b6253-61b6-49b5-99ee-40376865e0bd