Improve "create artifact" scroll.
Now only generates amulets/rings, books, potions or scrolls. In theory,
weapons and armour could be interesting as well, but since you can buy
all of them in the shop that'd be a bit pointless. We've got a similar
problem with the rings, but we can't be sure the scroll will create an
amulet. The generated item is guaranteed not to be cursed or negatively
enchanted.

When blessed, may create amulets even on low levels, potions/scrolls
are guaranteed to not be a dud item, and books are more likely to be 
unknown and of a higher level.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@660 561b6253-61b6-49b5-99ee-40376865e0bd