Populate the town.
Add a new monster type, town person, that won't ever attack the player.
Four of these are placed in the town at game start but they get a new
random name once in a while to create the illusion of a lively town.
No gameplay change, this is pure flavour.

When you are adjacent to one of them, they will spout a random fortune.
It is surprisingly fun to walk about town and talk to those guys. Even
more astoundingly, I now feel we need more fortunes. :)


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@663 561b6253-61b6-49b5-99ee-40376865e0bd