Don't add non-unique weapons sold by the player to the store.
This greatly reduces the clutter even if the players sells anything he
finds. All weapon types are already available in the shop from the
beginning, and if you somehow manage to accidentally sell your best
weapon, you can either get a new one (mundane weapons are common) or
it's an artefact, in which case the shop will gladly take it back.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@668 561b6253-61b6-49b5-99ee-40376865e0bd