Fix nymphs causing crashes when stealing equipped items.
Now that some actions are interruptible, an action may sometimes try to 
interrupt itself. When this happens when a nymph attempts to steal an
equipped item and while you're busy unequipping said item she goes for
the same item again, it won't be equipped anymore the second time around
and the crashes when trying to access the equipment slot pointer, eg.
eq_weapon. This is more likely to happen when you carry few items or
have lots of equipment slots filled.

This fix checks whether the pointer is still valid before trying to
access it. A saner solution would involve making sure that an action
doesn't trigger an interruption of itself.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@672 561b6253-61b6-49b5-99ee-40376865e0bd