Add rivers and lakes, yay! :D
Code copied from Dungeon Crawl (Stonesoup), hope that's okay.

Spells can now be cast across water and lava, and flying monsters can 
cross it. If a monster is killed over water/lava, all carried items get
destroyed - except for the pcd, eye of larn and weapon artefacts which
get teleported elsewhere on the level. Um, it's magic?

Also fixes walk through walls not working when blind, and the message
when bumping into water/lava. (I opted against allowing the player to
drown this way.)

The chances of rivers/lakes actually appearing are much lower than the
base probability in map.c (90% lava in the volcano, 40% water in the 
caverns) because the connectivity check may cause the map to vetoed 
and create a new one.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@703 561b6253-61b6-49b5-99ee-40376865e0bd