Some modifications of the custom maze maps.
The amulet of larn (previously '*') and potion of cure dianthroritis are
now both depicted by '!', of which there are three instances per map
now, one of which is chosen at random. 
I really didn't like the fact that once you recognize the map layout 
(and some of them are very recognizable) it was possible to completely 
predict the item position. Not that I ever did that, but I do think that
it shouldn't be this easy.

Random items (previously '-') are now depicted by 'o', which is easier
to see.


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@715 561b6253-61b6-49b5-99ee-40376865e0bd