Reduced availability of items in the store.
Johanna explained the reasons in a ticket:
(see https://sourceforge.net/apps/trac/nlarn/ticket/250)

For reasons of preventing start-scumming, the shop should always stock
the same set of items. However, I do feel that it currently guarantees
too many items, which a) reduces the impact of finding items randomly,
and b) plays a large role in making all games feel the same.

1.) Rings
Currently, all rings are available in the shop. I think it might be a
good idea to limit those to the early stat-boosting rings, i.e.
protection, strength, dexterity and intelligence.

2.) Potions and scrolls
Potions and scrolls are already only a subset of all available items,
but I don't think it's necessary to stock all of them in a quantity of
3. Changing 'obtainable' from a boolean to an int would allow
distinction between common items (healing, enchant foo) and less common
ones (heroism, enlightenment).

Even so, cutting a few more of them entirely might be worth it, for
example spirit protection or treasure finding.

3.) Books
Books are already highly restricted, but more could be cut. In my
opinion, the ones that absolutely must be kept, are the curation ones
(blindness and poison) as well as one attack spell, for which I'd
suggest magic missile. For variation, I'd also include protection, sleep
and possibly healing, but IMHO all others could be removed from the shop
without all hell breaking lose.

4.) Armour
Upgrading weapons early by way of the shop seems a viable strategy at
higher difficulty, but I don't think we really need to offer absolutely
all types of armour. In my opinion, restricting the selection to the
first five base items for each slot and a couple of upgrades for the
body armour (say, ring mail and plate mail) wouldn't really harm the
game.