Ensure player's experience can not sink below zero
One of the effect of quaffing from fountains is that experience may be
lowered. When experience is low, it is possible that the player is about
to lose more experience that he has, so the unsigned integer used to
store the player's experience points may wrap around and all hell may
break loose (i.e. segmentation fault).