Increase max hp/mp whe using restoring potions.
Increase the hp/mp max values when using the potion of healing, the
potion of instant healing and the potion of power. Increase by 1% for
the "small" potions, 5% for the "big" potion (currently there is no
potion of MAX_MP).

Positive side effect: display a message when using the potion of full
healing.