RNG redesign
Exchange the Mersenne Twister PRPG provided by GLib with
xoshiro128**. This allows us to save the state of the PRNG
with the game,thus ensuring the game will always continue
identically when restoring it two times.

As a nice side effect, the new PRNG is significantly faster,
reducing the time required to create the game.