Revert "Define colours enum with enum factory"
This reverts commit 0e3f8d4c169181330b42989969645bba2e701359.

This commit caused issues with spell animations, as defining
the enum value for no colour with -1 to avoid the duplicate index
meddled with the deep internals of curses. Keeping it as 0 didn't
work either, as light gray is also defined as 0 (don't ask me why,
that's just the way it is).