Ensure items on player's position are not mentioned twice
Now that we log all items on the player's position, some items might get
logged twice, if they were configured for auto-pickup but to heavy to
get. Those were logged as "The 3896 gold are too heavy for you. You ser
3896 gold here.". This change ensures that items of types which are
configured for auto-pickup are not mentioned in the list of items on the
player's position.