Switch to 256 colors
This became necessary as modern terminals no longer interpret A_BOLD as
bright colour, hence the display of this game was broken in some
terminals in their default configuration. I've noticed this in the new
Terminal included in Fedora 41, but also the sweet COSMIC Terminal behaves
like this. With 256 colours, we do not need to use A_BOLD to modify
colours anymore and the first-time experience is okay gain.

Unfortunately PDCurses only supports 256 color pairs, so we cannot use
all possible combinations of foreground and background colours.

This way the colour names also serve as color pair names. The 256 colour
palette contains 8 duplicate colour definitions, which we can use for
pairs with different background colour, most notably the UI colours.

The colour names were carefully assembled from various sources, but most
notable Thomas Dickey's helper script closest-rgb.pl [^1], which was
designed to find the closest colour names for the XTerm colour palette.

This script was run against the original X11 list of colours [^2],
Wikipedia's list of RAL [^3], Crayola crayon colors [^4] and generic
colour names [^5], and the list of traditional Japanese colour names,
which can be found on Wikipedia [^6], but the data was taken from a
JavaScript representation of said data [^7].

The script not only shows the closest match for a value, but also the
distance of the colour to the named colour. Usually I've picked the
colour name with the best match, but there were quite a few duplicates
and bad matches. Those I've fed into the ArtyClick Color Name Finder
[^8], which knew a nice name for most of the remaining colours.

I have assigned new colours to most items and monsters. For the latter
I've tried to assign a colour that matches the well-known D&D style of
the monsters.

[^1]: https://github.com/ThomasDickey/xterm-snapshots/blob/master/vttests/closest-rgb.pl
[^2]: https://gitlab.freedesktop.org/xorg/app/rgb/-/blob/master/rgb.txt
[^3]: https://en.wikipedia.org/wiki/List_of_RAL_colours
[^4]: https://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors
[^5]: https://en.wikipedia.org/wiki/List_of_colors:_A%E2%80%93F
      https://en.wikipedia.org/wiki/List_of_colors:_G%E2%80%93M
      https://en.wikipedia.org/wiki/List_of_colors:_N%E2%80%93Z
[^6]: https://en.wikipedia.org/wiki/Traditional_colors_of_Japan
[^7]: https://gist.github.com/jpegzilla/2ab93a895b0e484fa042b7bde29a093c
[^8]: https://colors.artyclick.com/color-name-finder/